% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_distchange.R
\name{calc_distchange}
\alias{calc_distchange}
\title{Calculate distance change in future distribution}
\usage{
calc_distchange(
  eachspecies,
  sdm.species,
  crs.eqap,
  sdm.names,
  sdm.files,
  present.name,
  sdm.scenarios
)
}
\arguments{
\item{eachspecies}{}

\item{sdm.species}{}

\item{crs.eqap}{}

\item{sdm.names}{}

\item{sdm.files}{}

\item{present.name}{}

\item{sdm.scenarios}{}
}
\description{
Calculate distance change in future distribution
}
