% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-comunidadescosteras.R
\docType{data}
\name{comunidadescosteras}
\alias{comunidadescosteras}
\title{DATASET_TITLE}
\format{
A data frame with 12905 rows and 239 variables:
\describe{
\item{\code{COM_ID}}{character COLUMN_DESCRIPTION}
\item{\code{ENTIDAD}}{double COLUMN_DESCRIPTION}
\item{\code{NOM_ENT}}{character COLUMN_DESCRIPTION}
\item{\code{MUN}}{double COLUMN_DESCRIPTION}
\item{\code{NOM_MUN}}{character COLUMN_DESCRIPTION}
\item{\code{LOC}}{double COLUMN_DESCRIPTION}
\item{\code{NOM_LOC}}{character COLUMN_DESCRIPTION}
\item{\code{lon_deg}}{double COLUMN_DESCRIPTION}
\item{\code{lon_min}}{double COLUMN_DESCRIPTION}
\item{\code{lon_sec}}{double COLUMN_DESCRIPTION}
\item{\code{lat_deg}}{double COLUMN_DESCRIPTION}
\item{\code{lat_min}}{double COLUMN_DESCRIPTION}
\item{\code{lat_sec}}{double COLUMN_DESCRIPTION}
\item{\code{ALTITUD}}{double COLUMN_DESCRIPTION}
\item{\code{POBTOT}}{double COLUMN_DESCRIPTION}
\item{\code{POBFEM}}{double COLUMN_DESCRIPTION}
\item{\code{POBMAS}}{double COLUMN_DESCRIPTION}
\item{\code{P_0A2}}{double COLUMN_DESCRIPTION}
\item{\code{P_0A2_F}}{double COLUMN_DESCRIPTION}
\item{\code{P_0A2_M}}{double COLUMN_DESCRIPTION}
\item{\code{P_3YMAS}}{double COLUMN_DESCRIPTION}
\item{\code{P_3YMAS_F}}{double COLUMN_DESCRIPTION}
\item{\code{P_3YMAS_M}}{double COLUMN_DESCRIPTION}
\item{\code{P_5YMAS}}{double COLUMN_DESCRIPTION}
\item{\code{P_5YMAS_F}}{double COLUMN_DESCRIPTION}
\item{\code{P_5YMAS_M}}{double COLUMN_DESCRIPTION}
\item{\code{P_12YMAS}}{double COLUMN_DESCRIPTION}
\item{\code{P_12YMAS_F}}{double COLUMN_DESCRIPTION}
\item{\code{P_12YMAS_M}}{double COLUMN_DESCRIPTION}
\item{\code{P_15YMAS}}{double COLUMN_DESCRIPTION}
\item{\code{P_15YMAS_F}}{double COLUMN_DESCRIPTION}
\item{\code{P_15YMAS_M}}{double COLUMN_DESCRIPTION}
\item{\code{P_18YMAS}}{double COLUMN_DESCRIPTION}
\item{\code{P_18YMAS_F}}{double COLUMN_DESCRIPTION}
\item{\code{P_18YMAS_M}}{double COLUMN_DESCRIPTION}
\item{\code{P_3A5}}{double COLUMN_DESCRIPTION}
\item{\code{P_3A5_F}}{double COLUMN_DESCRIPTION}
\item{\code{P_3A5_M}}{double COLUMN_DESCRIPTION}
\item{\code{P_6A11}}{double COLUMN_DESCRIPTION}
\item{\code{P_6A11_F}}{double COLUMN_DESCRIPTION}
\item{\code{P_6A11_M}}{double COLUMN_DESCRIPTION}
\item{\code{P_8A14}}{double COLUMN_DESCRIPTION}
\item{\code{P_8A14_F}}{double COLUMN_DESCRIPTION}
\item{\code{P_8A14_M}}{double COLUMN_DESCRIPTION}
\item{\code{P_12A14}}{double COLUMN_DESCRIPTION}
\item{\code{P_12A14_F}}{double COLUMN_DESCRIPTION}
\item{\code{P_12A14_M}}{double COLUMN_DESCRIPTION}
\item{\code{P_15A17}}{double COLUMN_DESCRIPTION}
\item{\code{P_15A17_F}}{double COLUMN_DESCRIPTION}
\item{\code{P_15A17_M}}{double COLUMN_DESCRIPTION}
\item{\code{P_18A24}}{double COLUMN_DESCRIPTION}
\item{\code{P_18A24_F}}{double COLUMN_DESCRIPTION}
\item{\code{P_18A24_M}}{double COLUMN_DESCRIPTION}
\item{\code{P_15A49_F}}{double COLUMN_DESCRIPTION}
\item{\code{P_60YMAS}}{double COLUMN_DESCRIPTION}
\item{\code{P_60YMAS_F}}{double COLUMN_DESCRIPTION}
\item{\code{P_60YMAS_M}}{double COLUMN_DESCRIPTION}
\item{\code{REL_H_M}}{double COLUMN_DESCRIPTION}
\item{\code{POB0_14}}{double COLUMN_DESCRIPTION}
\item{\code{POB15_64}}{double COLUMN_DESCRIPTION}
\item{\code{POB65_MAS}}{double COLUMN_DESCRIPTION}
\item{\code{PROM_HNV}}{double COLUMN_DESCRIPTION}
\item{\code{PNACENT}}{double COLUMN_DESCRIPTION}
\item{\code{PNACENT_F}}{double COLUMN_DESCRIPTION}
\item{\code{PNACENT_M}}{double COLUMN_DESCRIPTION}
\item{\code{PNACOE}}{double COLUMN_DESCRIPTION}
\item{\code{PNACOE_F}}{double COLUMN_DESCRIPTION}
\item{\code{PNACOE_M}}{double COLUMN_DESCRIPTION}
\item{\code{PRES2015}}{double COLUMN_DESCRIPTION}
\item{\code{PRES2015_F}}{double COLUMN_DESCRIPTION}
\item{\code{PRES2015_M}}{double COLUMN_DESCRIPTION}
\item{\code{PRESOE15}}{double COLUMN_DESCRIPTION}
\item{\code{PRESOE15_F}}{double COLUMN_DESCRIPTION}
\item{\code{PRESOE15_M}}{double COLUMN_DESCRIPTION}
\item{\code{P3YM_HLI}}{double COLUMN_DESCRIPTION}
\item{\code{P3YM_HLI_F}}{double COLUMN_DESCRIPTION}
\item{\code{P3YM_HLI_M}}{double COLUMN_DESCRIPTION}
\item{\code{P3HLINHE}}{double COLUMN_DESCRIPTION}
\item{\code{P3HLINHE_F}}{double COLUMN_DESCRIPTION}
\item{\code{P3HLINHE_M}}{double COLUMN_DESCRIPTION}
\item{\code{P3HLI_HE}}{double COLUMN_DESCRIPTION}
\item{\code{P3HLI_HE_F}}{double COLUMN_DESCRIPTION}
\item{\code{P3HLI_HE_M}}{double COLUMN_DESCRIPTION}
\item{\code{P5_HLI}}{double COLUMN_DESCRIPTION}
\item{\code{P5_HLI_NHE}}{double COLUMN_DESCRIPTION}
\item{\code{P5_HLI_HE}}{double COLUMN_DESCRIPTION}
\item{\code{PHOG_IND}}{double COLUMN_DESCRIPTION}
\item{\code{POB_AFRO}}{double COLUMN_DESCRIPTION}
\item{\code{POB_AFRO_F}}{double COLUMN_DESCRIPTION}
\item{\code{POB_AFRO_M}}{double COLUMN_DESCRIPTION}
\item{\code{PCON_DISC}}{double COLUMN_DESCRIPTION}
\item{\code{PCDISC_MOT}}{double COLUMN_DESCRIPTION}
\item{\code{PCDISC_VIS}}{double COLUMN_DESCRIPTION}
\item{\code{PCDISC_LENG}}{double COLUMN_DESCRIPTION}
\item{\code{PCDISC_AUD}}{double COLUMN_DESCRIPTION}
\item{\code{PCDISC_MOT2}}{double COLUMN_DESCRIPTION}
\item{\code{PCDISC_MEN}}{double COLUMN_DESCRIPTION}
\item{\code{PCON_LIMI}}{double COLUMN_DESCRIPTION}
\item{\code{PCLIM_CSB}}{double COLUMN_DESCRIPTION}
\item{\code{PCLIM_VIS}}{double COLUMN_DESCRIPTION}
\item{\code{PCLIM_HACO}}{double COLUMN_DESCRIPTION}
\item{\code{PCLIM_OAUD}}{double COLUMN_DESCRIPTION}
\item{\code{PCLIM_MOT2}}{double COLUMN_DESCRIPTION}
\item{\code{PCLIM_RE_CO}}{double COLUMN_DESCRIPTION}
\item{\code{PCLIM_PMEN}}{double COLUMN_DESCRIPTION}
\item{\code{PSIND_LIM}}{double COLUMN_DESCRIPTION}
\item{\code{P3A5_NOA}}{double COLUMN_DESCRIPTION}
\item{\code{P3A5_NOA_F}}{double COLUMN_DESCRIPTION}
\item{\code{P3A5_NOA_M}}{double COLUMN_DESCRIPTION}
\item{\code{P6A11_NOA}}{double COLUMN_DESCRIPTION}
\item{\code{P6A11_NOAF}}{double COLUMN_DESCRIPTION}
\item{\code{P6A11_NOAM}}{double COLUMN_DESCRIPTION}
\item{\code{P12A14NOA}}{double COLUMN_DESCRIPTION}
\item{\code{P12A14NOAF}}{double COLUMN_DESCRIPTION}
\item{\code{P12A14NOAM}}{double COLUMN_DESCRIPTION}
\item{\code{P15A17A}}{double COLUMN_DESCRIPTION}
\item{\code{P15A17A_F}}{double COLUMN_DESCRIPTION}
\item{\code{P15A17A_M}}{double COLUMN_DESCRIPTION}
\item{\code{P18A24A}}{double COLUMN_DESCRIPTION}
\item{\code{P18A24A_F}}{double COLUMN_DESCRIPTION}
\item{\code{P18A24A_M}}{double COLUMN_DESCRIPTION}
\item{\code{P8A14AN}}{double COLUMN_DESCRIPTION}
\item{\code{P8A14AN_F}}{double COLUMN_DESCRIPTION}
\item{\code{P8A14AN_M}}{double COLUMN_DESCRIPTION}
\item{\code{P15YM_AN}}{double COLUMN_DESCRIPTION}
\item{\code{P15YM_AN_F}}{double COLUMN_DESCRIPTION}
\item{\code{P15YM_AN_M}}{double COLUMN_DESCRIPTION}
\item{\code{P15YM_SE}}{double COLUMN_DESCRIPTION}
\item{\code{P15YM_SE_F}}{double COLUMN_DESCRIPTION}
\item{\code{P15YM_SE_M}}{double COLUMN_DESCRIPTION}
\item{\code{P15PRI_IN}}{double COLUMN_DESCRIPTION}
\item{\code{P15PRI_INF}}{double COLUMN_DESCRIPTION}
\item{\code{P15PRI_INM}}{double COLUMN_DESCRIPTION}
\item{\code{P15PRI_CO}}{double COLUMN_DESCRIPTION}
\item{\code{P15PRI_COF}}{double COLUMN_DESCRIPTION}
\item{\code{P15PRI_COM}}{double COLUMN_DESCRIPTION}
\item{\code{P15SEC_IN}}{double COLUMN_DESCRIPTION}
\item{\code{P15SEC_INF}}{double COLUMN_DESCRIPTION}
\item{\code{P15SEC_INM}}{double COLUMN_DESCRIPTION}
\item{\code{P15SEC_CO}}{double COLUMN_DESCRIPTION}
\item{\code{P15SEC_COF}}{double COLUMN_DESCRIPTION}
\item{\code{P15SEC_COM}}{double COLUMN_DESCRIPTION}
\item{\code{P18YM_PB}}{double COLUMN_DESCRIPTION}
\item{\code{P18YM_PB_F}}{double COLUMN_DESCRIPTION}
\item{\code{P18YM_PB_M}}{double COLUMN_DESCRIPTION}
\item{\code{GRAPROES}}{double COLUMN_DESCRIPTION}
\item{\code{GRAPROES_F}}{double COLUMN_DESCRIPTION}
\item{\code{GRAPROES_M}}{double COLUMN_DESCRIPTION}
\item{\code{PEA}}{double COLUMN_DESCRIPTION}
\item{\code{PEA_F}}{double COLUMN_DESCRIPTION}
\item{\code{PEA_M}}{double COLUMN_DESCRIPTION}
\item{\code{PE_INAC}}{double COLUMN_DESCRIPTION}
\item{\code{PE_INAC_F}}{double COLUMN_DESCRIPTION}
\item{\code{PE_INAC_M}}{double COLUMN_DESCRIPTION}
\item{\code{POCUPADA}}{double COLUMN_DESCRIPTION}
\item{\code{POCUPADA_F}}{double COLUMN_DESCRIPTION}
\item{\code{POCUPADA_M}}{double COLUMN_DESCRIPTION}
\item{\code{PDESOCUP}}{double COLUMN_DESCRIPTION}
\item{\code{PDESOCUP_F}}{double COLUMN_DESCRIPTION}
\item{\code{PDESOCUP_M}}{double COLUMN_DESCRIPTION}
\item{\code{PSINDER}}{double COLUMN_DESCRIPTION}
\item{\code{PDER_SS}}{double COLUMN_DESCRIPTION}
\item{\code{PDER_IMSS}}{double COLUMN_DESCRIPTION}
\item{\code{PDER_ISTE}}{double COLUMN_DESCRIPTION}
\item{\code{PDER_ISTEE}}{double COLUMN_DESCRIPTION}
\item{\code{PAFIL_PDOM}}{double COLUMN_DESCRIPTION}
\item{\code{PDER_SEGP}}{double COLUMN_DESCRIPTION}
\item{\code{PDER_IMSSB}}{double COLUMN_DESCRIPTION}
\item{\code{PAFIL_IPRIV}}{double COLUMN_DESCRIPTION}
\item{\code{PAFIL_OTRAI}}{double COLUMN_DESCRIPTION}
\item{\code{P12YM_SOLT}}{double COLUMN_DESCRIPTION}
\item{\code{P12YM_CASA}}{double COLUMN_DESCRIPTION}
\item{\code{P12YM_SEPA}}{double COLUMN_DESCRIPTION}
\item{\code{PCATOLICA}}{double COLUMN_DESCRIPTION}
\item{\code{PRO_CRIEVA}}{double COLUMN_DESCRIPTION}
\item{\code{POTRAS_REL}}{double COLUMN_DESCRIPTION}
\item{\code{PSIN_RELIG}}{double COLUMN_DESCRIPTION}
\item{\code{TOTHOG}}{double COLUMN_DESCRIPTION}
\item{\code{HOGJEF_F}}{double COLUMN_DESCRIPTION}
\item{\code{HOGJEF_M}}{double COLUMN_DESCRIPTION}
\item{\code{POBHOG}}{double COLUMN_DESCRIPTION}
\item{\code{PHOGJEF_F}}{double COLUMN_DESCRIPTION}
\item{\code{PHOGJEF_M}}{double COLUMN_DESCRIPTION}
\item{\code{VIVTOT}}{double COLUMN_DESCRIPTION}
\item{\code{TVIVHAB}}{double COLUMN_DESCRIPTION}
\item{\code{TVIVPAR}}{double COLUMN_DESCRIPTION}
\item{\code{VIVPAR_HAB}}{double COLUMN_DESCRIPTION}
\item{\code{TVIVPARHAB}}{double COLUMN_DESCRIPTION}
\item{\code{VIVPAR_DES}}{double COLUMN_DESCRIPTION}
\item{\code{VIVPAR_UT}}{double COLUMN_DESCRIPTION}
\item{\code{OCUPVIVPAR}}{double COLUMN_DESCRIPTION}
\item{\code{PROM_OCUP}}{double COLUMN_DESCRIPTION}
\item{\code{PRO_OCUP_C}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_PISODT}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_PISOTI}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_1DOR}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_2YMASD}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_1CUART}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_2CUART}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_3YMASC}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_C_ELEC}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_S_ELEC}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_AGUADV}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_AEASP}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_AGUAFV}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_TINACO}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_CISTER}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_EXCSA}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_LETR}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_DRENAJ}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_NODREN}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_C_SERV}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_NDEAED}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_DSADMA}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_NDACMM}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_SNBIEN}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_REFRI}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_LAVAD}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_HMICRO}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_AUTOM}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_MOTO}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_BICI}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_RADIO}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_TV}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_PC}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_TELEF}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_CEL}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_INTER}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_STVP}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_SPMVPI}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_CVJ}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_SINRTV}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_SINLTC}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_SINCINT}}{double COLUMN_DESCRIPTION}
\item{\code{VPH_SINTIC}}{double COLUMN_DESCRIPTION}
\item{\code{TAMLOC}}{double COLUMN_DESCRIPTION}
\item{\code{id_row}}{double COLUMN_DESCRIPTION}
\item{\code{dec_lon}}{double COLUMN_DESCRIPTION}
\item{\code{dec_lat}}{double COLUMN_DESCRIPTION}
}
}
\usage{
comunidadescosteras
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
