% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctcomid.R
\name{correctcomid}
\alias{correctcomid}
\title{Correct COM_ID field to CVE_LOC}
\usage{
correctcomid(thiscomponentdata)
}
\arguments{
\item{thiscomponentdata}{}
}
\description{
Correct COM_ID field to CVE_LOC
}
