% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-exposuressp585.R
\docType{data}
\name{exposuressp585}
\alias{exposuressp585}
\title{DATASET_TITLE}
\format{
A data frame with 2219 rows and 35 variables:
\describe{
\item{\code{CVE_MUN}}{character COLUMN_DESCRIPTION}
\item{\code{CVE_LOC}}{character COLUMN_DESCRIPTION}
\item{\code{Average_Chlorophyll}}{double COLUMN_DESCRIPTION}
\item{\code{Average_DissolvedIron}}{double COLUMN_DESCRIPTION}
\item{\code{Average_DissolvedMolecularOxygen}}{double COLUMN_DESCRIPTION}
\item{\code{Average_Nitrate}}{double COLUMN_DESCRIPTION}
\item{\code{Average_OceanTemperature}}{double COLUMN_DESCRIPTION}
\item{\code{Average_pH}}{double COLUMN_DESCRIPTION}
\item{\code{Average_Salinity}}{double COLUMN_DESCRIPTION}
\item{\code{Average_SeaWaterDirection}}{double COLUMN_DESCRIPTION}
\item{\code{Average_SeaWaterSpeed}}{double COLUMN_DESCRIPTION}
\item{\code{Average_Silicate}}{double COLUMN_DESCRIPTION}
\item{\code{Average_TotalPhytoplankton}}{double COLUMN_DESCRIPTION}
\item{\code{Maximum_Chlorophyll}}{double COLUMN_DESCRIPTION}
\item{\code{Maximum_DissolvedIron}}{double COLUMN_DESCRIPTION}
\item{\code{Maximum_DissolvedMolecularOxygen}}{double COLUMN_DESCRIPTION}
\item{\code{Maximum_Nitrate}}{double COLUMN_DESCRIPTION}
\item{\code{Maximum_OceanTemperature}}{double COLUMN_DESCRIPTION}
\item{\code{Maximum_pH}}{double COLUMN_DESCRIPTION}
\item{\code{Maximum_Salinity}}{double COLUMN_DESCRIPTION}
\item{\code{Maximum_SeaWaterDirection}}{double COLUMN_DESCRIPTION}
\item{\code{Maximum_SeaWaterSpeed}}{double COLUMN_DESCRIPTION}
\item{\code{Maximum_Silicate}}{double COLUMN_DESCRIPTION}
\item{\code{Maximum_TotalPhytoplankton}}{double COLUMN_DESCRIPTION}
\item{\code{Minimum_Chlorophyll}}{double COLUMN_DESCRIPTION}
\item{\code{Minimum_DissolvedIron}}{double COLUMN_DESCRIPTION}
\item{\code{Minimum_DissolvedMolecularOxygen}}{double COLUMN_DESCRIPTION}
\item{\code{Minimum_Nitrate}}{double COLUMN_DESCRIPTION}
\item{\code{Minimum_OceanTemperature}}{double COLUMN_DESCRIPTION}
\item{\code{Minimum_pH}}{double COLUMN_DESCRIPTION}
\item{\code{Minimum_Salinity}}{double COLUMN_DESCRIPTION}
\item{\code{Minimum_SeaWaterDirection}}{double COLUMN_DESCRIPTION}
\item{\code{Minimum_SeaWaterSpeed}}{double COLUMN_DESCRIPTION}
\item{\code{Minimum_Silicate}}{double COLUMN_DESCRIPTION}
\item{\code{Minimum_TotalPhytoplankton}}{double COLUMN_DESCRIPTION}
}
}
\usage{
exposuressp585
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
