% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-localities.R
\docType{data}
\name{localities}
\alias{localities}
\title{DATASET_TITLE}
\format{
A data frame with 118 rows and 10 variables:
\describe{
\item{\code{fip_name}}{character COLUMN_DESCRIPTION}
\item{\code{fip_id_number}}{integer COLUMN_DESCRIPTION}
\item{\code{ENT}}{integer COLUMN_DESCRIPTION}
\item{\code{MUN}}{integer COLUMN_DESCRIPTION}
\item{\code{LOC}}{integer COLUMN_DESCRIPTION}
\item{\code{COM_ID}}{character COLUMN_DESCRIPTION}
\item{\code{locality}}{character COLUMN_DESCRIPTION}
\item{\code{municipality}}{character COLUMN_DESCRIPTION}
\item{\code{state}}{character COLUMN_DESCRIPTION}
\item{\code{participants}}{character COLUMN_DESCRIPTION}
}
}
\usage{
localities
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
