% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_var_expo.R
\name{scale_var_expo}
\alias{scale_var_expo}
\title{Scale variable}
\usage{
scale_var_expo(datatable, sspscenario, vars.expo, fiplocal.id)
}
\arguments{
\item{sspscenario}{}
}
\value{
num.scale data with scaled total
}
\description{
Scale variable
}
\examples{
scaled.exp <- scale_var(exp.data.set.sc, NA)
}
