% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sensitivitydata.R
\docType{data}
\name{sensitivitydata}
\alias{sensitivitydata}
\title{DATASET_TITLE}
\format{
A data frame with 70 rows and 34 variables:
\describe{
\item{\code{CVE_ENT}}{character COLUMN_DESCRIPTION}
\item{\code{CVE_LOC}}{character COLUMN_DESCRIPTION}
\item{\code{CVE_MUN}}{character COLUMN_DESCRIPTION}
\item{\code{COM_MUN}}{character COLUMN_DESCRIPTION}
\item{\code{POBTOT}}{double COLUMN_DESCRIPTION}
\item{\code{tot_captura_loc}}{double COLUMN_DESCRIPTION}
\item{\code{tot_valor_loc}}{double COLUMN_DESCRIPTION}
\item{\code{tot_captura_mun}}{double COLUMN_DESCRIPTION}
\item{\code{tot_valor_mun}}{double COLUMN_DESCRIPTION}
\item{\code{pob_mun}}{double COLUMN_DESCRIPTION}
\item{\code{tot_captura_mun_per_capita}}{double COLUMN_DESCRIPTION}
\item{\code{tot_captura_loc_per_capita}}{double COLUMN_DESCRIPTION}
\item{\code{tot_valor_loc_per_capita}}{double COLUMN_DESCRIPTION}
\item{\code{ENTIDAD}}{double COLUMN_DESCRIPTION}
\item{\code{NOM_ENT}}{character COLUMN_DESCRIPTION}
\item{\code{MUN}}{double COLUMN_DESCRIPTION}
\item{\code{NOM_MUN}}{character COLUMN_DESCRIPTION}
\item{\code{LOC}}{double COLUMN_DESCRIPTION}
\item{\code{NOM_LOC}}{character COLUMN_DESCRIPTION}
\item{\code{num_permisos_pesca_mun}}{double COLUMN_DESCRIPTION}
\item{\code{tot_pob_mun}}{double COLUMN_DESCRIPTION}
\item{\code{permisos_pesca_per_capita_mun}}{double COLUMN_DESCRIPTION}
\item{\code{num_permisos_acuicolas}}{double COLUMN_DESCRIPTION}
\item{\code{permisos_acui_per_capita_mun}}{double COLUMN_DESCRIPTION}
\item{\code{neg_acuicolas}}{double COLUMN_DESCRIPTION}
\item{\code{neg_pesca}}{double COLUMN_DESCRIPTION}
\item{\code{tot_ue}}{double COLUMN_DESCRIPTION}
\item{\code{prop_acuicolas}}{double COLUMN_DESCRIPTION}
\item{\code{prop_pesca}}{double COLUMN_DESCRIPTION}
\item{\code{neg_acui_per_capita}}{double COLUMN_DESCRIPTION}
\item{\code{neg_pesca_per_capita}}{double COLUMN_DESCRIPTION}
\item{\code{no_loc}}{integer COLUMN_DESCRIPTION}
\item{\code{no_mun}}{integer COLUMN_DESCRIPTION}
\item{\code{no_ent}}{integer COLUMN_DESCRIPTION}
}
}
\usage{
sensitivitydata
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
