% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-specieslist.R
\docType{data}
\name{specieslist}
\alias{specieslist}
\title{DATASET_TITLE}
\format{
A data frame with 47 rows and 4 variables:
\describe{
\item{\code{taxa}}{character COLUMN_DESCRIPTION}
\item{\code{species}}{character COLUMN_DESCRIPTION}
\item{\code{common_name}}{character COLUMN_DESCRIPTION}
\item{\code{nombre_comun}}{character COLUMN_DESCRIPTION}
}
}
\usage{
specieslist
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
