% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-exposuressp585.R
\docType{data}
\name{exposuressp585}
\alias{exposuressp585}
\title{DATASET_TITLE}
\format{
A data frame with 70 rows and 21 variables:
\describe{
\item{\code{CVE_ENT}}{character COLUMN_DESCRIPTION}
\item{\code{CVE_LOC}}{character COLUMN_DESCRIPTION}
\item{\code{CVE_MUN}}{character COLUMN_DESCRIPTION}
\item{\code{impacto_antropogenico}}{double COLUMN_DESCRIPTION}
\item{\code{impacto_humano}}{double COLUMN_DESCRIPTION}
\item{\code{intervalo_diurno}}{double COLUMN_DESCRIPTION}
\item{\code{marea_roja}}{double COLUMN_DESCRIPTION}
\item{\code{oxigeno_disuelto}}{double COLUMN_DESCRIPTION}
\item{\code{precipitacion_promedio}}{double COLUMN_DESCRIPTION}
\item{\code{productividad_primaria}}{double COLUMN_DESCRIPTION}
\item{\code{riesgo_huracanes}}{double COLUMN_DESCRIPTION}
\item{\code{riesgo_inundaciones}}{double COLUMN_DESCRIPTION}
\item{\code{temperatura_promedio_anual}}{double COLUMN_DESCRIPTION}
\item{\code{temperatura_superficial}}{double COLUMN_DESCRIPTION}
\item{\code{NOM_ENT}}{character COLUMN_DESCRIPTION}
\item{\code{NOM_MUN}}{character COLUMN_DESCRIPTION}
\item{\code{NOM_LOC}}{character COLUMN_DESCRIPTION}
\item{\code{POBTOT}}{double COLUMN_DESCRIPTION}
\item{\code{no_loc}}{integer COLUMN_DESCRIPTION}
\item{\code{no_mun}}{integer COLUMN_DESCRIPTION}
\item{\code{no_ent}}{integer COLUMN_DESCRIPTION}
}
}
\usage{
exposuressp585
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
